added aws base URL

+++ src/app/actors/all-actors/all-actors.component.ts
@@ -1,8 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { CommonModule } from '@angular/common';
-import { environment } from '../../../environments/environment.development';
+import { environment } from '../../../environments/environment';
 import { Actor } from '../actors';
+import { AuthService } from '../../auth/auth.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-all-actors',
@@ -14,7 +16,11 @@ import { Actor } from '../actors';
 export class AllActorsComponent implements OnInit {
   public actors: Actor[] = [];
 
-  constructor(private http: HttpClient) {}
+  constructor(
+    private http: HttpClient,
+    private router: Router,
+    private authService: AuthService
+  ) {}
 
   ngOnInit(): void {
     this.getAllActors();
@@ -31,6 +37,11 @@ export class AllActorsComponent implements OnInit {
     });
   }
   deleteActor(id: number): void {
+    if (!this.authService.isAuthenticated()) {
+      alert('You need to log in to delete movies.');
+      this.router.navigate(['/login']);
+      return;
+    }
     if (confirm('Are you sure you want to delete this actor?')) {
       this.http.delete(`${environment.baseUrl}api/Actors/${id}`).subscribe({
         next: () => {
